requirements.txt
flask
prometheus-client
protobuf
python-snappy
pyyaml
